% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avm.R
\name{avm.fit}
\alias{avm.fit}
\title{avm modeling}
\usage{
avm.fit(data.train = NULL, dependentY = "K0", independentX,
  method = "ols")
}
\arguments{
\item{data.train}{data for fitting generated by \code{\link{avm.data.train}} or \code{\link{data.split}}}

\item{dependentY}{dependent variable. The default is 'K0'}

\item{independentX}{independent variables. Using var exception,see detail.}

\item{method}{ols? rlm? rft? svm? xgb?}
}
\value{
The list of AVM models
}
\description{
model fitting function
}
