% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avm.R
\name{avm.pred.data}
\alias{avm.pred.data}
\title{generate data for prediction}
\usage{
avm.pred.data(train.model = NULL, tabln.vec = NULL, poi.data.list = NULL,
  method = "ols", proptype = "11", x = NULL, y = NULL,
  bldg_code = "11", byear = 5, height = 18, floor = 3, brArea = 90,
  stru = TRUE, face_through = TRUE, face_sun = TRUE, deco_if = TRUE,
  deco_good = TRUE, prop_rt = TRUE, byearmin = NULL, byearmax = NULL,
  heightmin = NULL, heightmax = NULL)
}
\arguments{
\item{train.model}{object of models provided by \code{\link{avm.fit}},required pameter.}

\item{tabln.vec}{basic data provided by \code{\link{loadData}}}

\item{poi.data.list}{data of POI ,required . It will be provided by \code{\link{getpoi.ras}}}

\item{method}{ols? rlm? svm? rft? xgb?}

\item{proptype}{type code of properties.The default is 11 which means housing}

\item{x}{lon,required}

\item{y}{lat,required}

\item{bldg_code}{tyep code of building.The default is 11}

\item{byear}{age of building}

\item{height}{total floors of building}

\item{floor}{which floor? The default is 3}

\item{brArea}{the area? The default is 90}

\item{stru}{logical}

\item{face_through}{logical}

\item{face_sun}{logical}

\item{deco_if}{logical}

\item{deco_good}{logical}

\item{prop_rt}{logical}

\item{byearmin}{the min byear of a community}

\item{byearmax}{the max byear of a community}

\item{heightmin}{the min height of a community}

\item{heightmax}{the max height of a community}
}
\value{
data used for predictiong
}
\description{
This function will generate data for given data
}
