% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avm_city.R
\name{avm_city}
\alias{avm_city}
\title{AVM modeling of each city}
\usage{
avm_city(city_code, proptype, model.list, host, port, user, password, dbname,
  modelpath1, modelpath2)
}
\arguments{
\item{city_code}{City's abbr}

\item{proptype}{Property type of house area, 11---residence, and else like offical area}

\item{model.list}{all models contain: ols, rlm, svm, rft, xgb}

\item{host}{Host of the server, character}

\item{port}{Port of the server, numeric}

\item{user}{User of the server, character}

\item{password}{Password of the server, character}

\item{dbname}{Database name of city, character}

\item{modelpath1}{The path storing models and tables with ha/poi information}

\item{modelpath2}{The path storing our final models, extending community models, e.g. community + city models}
}
\value{
write tif files
}
\description{
AVM modeling of each city
}
