% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avm.R
\name{formula_rvars}
\alias{formula_rvars}
\title{Get the valid dependent variable}
\usage{
formula_rvars(data = NULL,
  independentX = "~.-ha_code-K0-Kbyear-Kheight-Ktime-Kfloor-KbrArea-Kstru-Kface_through-Kface_sun-Kdeco_if-Kdeco_good-Kprop_rt",
  min.missing.rate = 0.4)
}
\arguments{
\item{data}{relatively rough data sets}

\item{independentX}{independent variables. Using var exception,see detail.}

\item{min.missing.rate}{The rate of missing rate for keeping, The default is 0.4 which indicting when the rate of a feature is more than 40 percent, it will be removed from data}
}
\value{
vector of dependent variable names
}
\description{
delete the independent variable that missing data is more than 40%
}
