% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avm.R
\name{user.pred}
\alias{user.pred}
\title{Prediction for user}
\usage{
user.pred(city_code = NULL, train.models = NULL, commodelpath = ".",
  tabln.vec = NULL, poi.data.list = NULL, proptype = "11", x = NULL,
  y = NULL, bldg_code = "11", byear = 3, height = 6, floor = 1,
  brArea = 90, stru = TRUE, face_through = TRUE, face_sun = TRUE,
  deco_if = TRUE, deco_good = TRUE, prop_rt = TRUE, byearmin = NULL,
  byearmax = NULL, heightmin = NULL, heightmax = NULL)
}
\arguments{
\item{city_code}{the city code}

\item{train.models}{object of models provided by \code{\link{avm.fit}},required pameter.}

\item{commodelpath}{the path storing community models or json files}

\item{tabln.vec}{basic data provided by \code{\link{loadData}}}

\item{poi.data.list}{data of POI ,required . It will be provided by \code{\link{getpoi.ras}}}

\item{proptype}{type code of properties.The default is 11 which means housing}

\item{x}{lon,required}

\item{y}{lat,required}

\item{bldg_code}{tyep code of building.The default is 11}

\item{byear}{age of building}

\item{height}{total floors of building}

\item{floor}{which floor? The default is 3}

\item{brArea}{the area? The default is 90}

\item{stru}{logical}

\item{face_through}{logical}

\item{face_sun}{logical}

\item{deco_if}{logical}

\item{deco_good}{logical}

\item{byearmin}{the min byear of a community}

\item{byearmax}{the max byear of a community}

\item{heightmin}{the min height of a community}

\item{heightmax}{the max height of a community}

\item{decoprop_rt}{logical}
}
\value{
data used for predictiong
}
\description{
This function is used for prediction by the user
}
